feat!: configure Discard Bot for sending a simple message.
Create an instance of the Discord Client, using 2 collback function. On_ready () event is called when the bot has
finished logging in and configuring; on_message() event is called when the bot has received a message.
When user type "!hello", the bot send an "Hello there! message. This configuration is use to test the
correct functioning of Discord Bot.